First commit
The idea with this project is to be able to generate symbols that look
somewhat like a face, and are thus recognizable to the viewer despite
not being anything close to an actual face. In it's current state, this
just randomly walks around and makes faces with absolutely no color
scheme at all.